requirements.txt
python-dotenv
setuptools>=80.9.0
wheel
python-multipart
httpx
fastapi
uvicorn[standard]
requests>=2.31
pydantic>=2.5
streamlit
chromadb==1.1.1